refactor: 등수 패키지 이동
